Make idt/gdt work without errors
The error lay in the fact that gcc optimized the large registers_t
struct to a pointer to that struct. This didn't work well with the
assembly that called the function since it was passing the struct by
reference on the stack. The solution was to push the stack pointer
itself and then chance the C function to one taking a pointer to
register_t.